First Kinda working state.  Needs to handle disconnections, also need to work on clock sync.

+++ DirectXProject/DirectXProject/Hardware.cpp
@@ -80,14 +80,14 @@ void HardwareState::Update()
 		loopCount++;
 		if (loopCount > loopMax) {
 			lastTime = currentTime;
-			currentTime = clock()/CLOCKS_PER_SEC;
+			currentTime = float(clock())/CLOCKS_PER_SEC;
 			timeForLastFrame = currentTime - lastTime;
 			timeForLastFrame /= loopCount;
 			loopCount = 0;
 		}
 	} else {
 		lastTime = currentTime;
-		currentTime = clock()/CLOCKS_PER_SEC + timerOffset;
+		currentTime = float(clock())/CLOCKS_PER_SEC + timerOffset;
 		timeForLastFrame = currentTime - lastTime;
 	}
 
@@ -125,7 +125,7 @@ void HardwareState::Update()
 
 double HardwareState::GetTimeSinceLastFrame()
 {
-	return clock()-currentTime;
+	return double(clock())-currentTime;
 }
 
 void HardwareState::OffsetTimer(double seconds)
